Doxygen for src/devices and src/cpu: rule 11.2 reaches 0
36 headers get about 430 blocks, and the functions that had no
hand-written declaration get one (types only; the parameter names live
in the @param lines): 137 in cpu_types.h, 31 in devices_types.h, plus
RTC_SetWallClockMode. The vendored ncr5386.[ch] and the generated
*_protos.h are untouched; no .c file changed for the documentation.

12 @brief entries say "unverified" where the code did not settle the
question, e.g. add_A_mem/sub_A_mem/rdiv_org/rmpy_org (the source carries
a fix-me note about C/O/Q), DMAControlBlocks_SetTXPointer (offset is
ignored), Modem_StartModem (role of address in server mode).

Types changed in definition AND declaration so the added declarations do
not introduce new rule 5.2 uses (same type on this platform):
signExtend short -> int16_t, MemoryWrite byte_select unsigned char ->
uint8_t, Device_IO_Seek offset long -> int64_t (cast at the fseek call),
cpu_set_ring_at_pf/_clpt long -> int64_t.

tools/house/audit.py: @param count ignores commas inside parentheses, so
a function-pointer parameter counts as one parameter. Three doc blocks
had been padded to match the old miscount; they describe the real
parameters again.

Proof: all 117 project objects identical to 5a6189b. Gate PASS incl.
TPE, Winchester, floppy, DAP. Rule 11.2: 853 -> 0.

+++ src/devices/device.c
@@ -318,14 +318,14 @@ void Device_SetInterruptStatus(Device *dev, bool active, uint16_t level)
     }
 }
 
-int32_t Device_IO_Seek(Device *dev, FILE *f, long offset)
+int32_t Device_IO_Seek(Device *dev, FILE *f, int64_t offset)
 {
     (void)dev;
     if (!f)
     {
         return -1;
     }
-    return fseek(f, offset, SEEK_SET);
+    return fseek(f, (long)offset, SEEK_SET); /* fseek takes long; images are far below 2 GB */
 }
